+++ Changelog.md
@@ -26,6 +26,10 @@ In general we introduce a few factories, that avoid having to pass the manifold 
   `MomentumGradient` works as before, but the manifold as its first parameter is no longer necessary and `p` is now a keyword argument.
 * `Nesterov` is now called `NesterovRule`.
   `Nesterov` works as before, but the manifold as its first parameter is no longer necessary and `p` is now a keyword argument.
+* `ConjugateDescentCoefficient` is now called `ConjugateDescentCoefficientRule`.
+  `ConjugateDescentCoefficient` works as before, but the manifold as its first parameter is no longer necessary.
+* `DaiYuanCoefficient` is now called `DaiYuanCoefficientRule`.
+  `DaiYuanCoefficient` works as before, but the manifold as its first parameter is no longer necessary.
 * `quasi_Newton` had a keyword `scale_initial_operator=` that was inconsistently declared (sometimes bool, sometimes real) and was unused.
   It is now called `initial_scale=1.0` and scales the initial (diagonal, unit) matrix within the approximation of the Hessian additionally to the $\frac{1}{\lVert g_k\rVert}$ scaling with the norm of the oldes gradient for the limited memory variant. For the full matrix variant the initial identity matrix is now scaled with this parameter.
 * Unify doc strings and presentation of keyword arguments